fixes and create tests

+++ src/main/java/com/kata/cinema/base/webapp/controllers/admin/AdminGenreRestController.java
@@ -1,8 +1,10 @@
 package com.kata.cinema.base.webapp.controllers.admin;
 
+import com.kata.cinema.base.exceptions.ExceptionOfId;
 import com.kata.cinema.base.models.dto.GenreResponseDto;
 import com.kata.cinema.base.models.entitys.Genres;
 import com.kata.cinema.base.service.abstracts.model.GenreService;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -14,6 +16,7 @@ import java.util.List;
 public class AdminGenreRestController {
     private final GenreService genreService;
 
+    @Autowired
     public AdminGenreRestController(GenreService genreService) {
         this.genreService = genreService;
     }
@@ -24,20 +27,29 @@ public class AdminGenreRestController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<GenreResponseDto> deleteGenres(@PathVariable Long id) {
-        genreService.deleteById(id);
-        return new ResponseEntity<>(HttpStatus.OK);
+    public ResponseEntity<GenreResponseDto> deleteGenres(@PathVariable Long id) throws ExceptionOfId {
+        if (genreService.isExistsById(id)) {
+            genreService.deleteById(id);
+            return new ResponseEntity<>(HttpStatus.OK);
+        } else {
+            throw new ExceptionOfId("There is no genre with this ID, try again.");
+        }
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<GenreResponseDto> update(@RequestBody Genres genres) {
+    public ResponseEntity<GenreResponseDto> update(@PathVariable Long id, @RequestParam String name) {
+        Genres genres = genreService.findById(id);
+        genres.setName(name);
         genreService.update(genres);
+        System.out.println(genreService.findGenreList());
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
-    @PostMapping
-    public ResponseEntity<GenreResponseDto> addGenre(@RequestBody Genres genres) {
-        genreService.save(genres);
+    @PostMapping()
+    public ResponseEntity<GenreResponseDto> addGenre(@RequestParam String name) {
+        genreService.save(name.transform(Genres::new));
+        System.out.println(genreService.findGenreList());
         return new ResponseEntity<>(HttpStatus.OK);
     }
+
 }
